# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Hisob24 API
  version: 0.1.0
  description: >
    Asosiy API kontrakti. Frontend uchun TS client shu fayldan generatsiya
    qilinadi (packages/api-client). Brauzer API'ga Next.js rewrites orqali
    /api prefiksi bilan murojaat qiladi. Boshqa saytdan kelgan o'zgartiruvchi
    so'rov (POST, PATCH, DELETE; Sec-Fetch-Site yoki Origin bo'yicha) 403
    forbidden oladi.
paths:
...
  /app/auth/sms/send:
    post:
      operationId: sendLoginCode
      summary: SMS bilan kirish kodini yuborish
      description: >
        Kod 6 xonali, 2 daqiqa amal qiladi. Tizimga faqat kamida bitta
        kompaniyaga a'zo user kiradi: boshqa raqamga SMS ketmaydi, lekin javob
        bir xil (kim borligi bilinmaydi). Bitta raqamga 60 soniyada bitta kod;
        IP'dan daqiqasiga 5 ta so'rov.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: "+998 90 123 45 67, 998901234567 yoki 901234567"
      responses:
        "200":
          description: Qabul qilindi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeSent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /app/auth/sms/verify:
    post:
      operationId: verifyLoginCode
      summary: SMS kod bilan kirish
      description: >
        Access token javobda (15 daqiqa), refresh token httpOnly
        refresh_token cookie'da (30 kun). User bitta company'da bo'lsa u
        tanlanadi, bir nechtasida company_id null. Kodga 5 ta xato urinish,
        IP'dan daqiqasiga 5 ta so'rov.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone:
                  type: string
                code:
                  type: string
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Kod noto'g'ri yoki muddati o'tgan (invalid_code)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /app/auth/telegram:
    post:
      operationId: telegramLogin
      summary: User Mini App'dan kodsiz kirish
      description: >
        initData USER_BOT_TOKEN bilan tekshiriladi (24 soat). Telegram akkaunt
        user botga yuborgan raqam kamida bitta kompaniyaga a'zo bo'lsa, SMS
        verify'dagi kabi token beriladi (bitta company bo'lsa tanlanadi). Bu
        sessiyaning
        refresh_token cookie'si Telegram Web iframe'ida ishlashi uchun
        SameSite=None; Secure; Partitioned (COOKIE_SECURE=true bo'lsa).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramLogin"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "401":
          description: initData yaroqsiz yoki eskirgan (invalid_init_data)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: >
            Raqam botga yuborilmagan (phone_not_shared) yoki raqam hech bir
            kompaniyaga a'zo emas (no_access, message raqamni aytadi)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /app/auth/refresh:
    post:
      operationId: refreshTokens
      summary: Access token'ni yangilash (rotation)
      description: >
        refresh_token cookie'si bekor qilinadi va yangisi beriladi; tanlangan
        company saqlanadi. A'zolik qayta tekshiriladi: user shu company'dan
        chiqarilgan bo'lsa, token company'siz beriladi; hech bir kompaniyaga
        a'zo bo'lmay qolgan bo'lsa, sessiya tugaydi (401).
      security:
        - refreshCookie: []
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "401":
          $ref: "#/components/responses/SessionEnded"
...
  /app/auth/switch-company:
    post:
      operationId: switchCompany
      summary: Company tanlash yoki tanlovni bekor qilish
      description: >
        Muddati o'tgan company'ga ham o'tish mumkin (uning sahifalari 402
        qaytaradi). company_id null tanlovni bekor qiladi: muddati o'tgan
        company'dan ro'yxatga qaytish yo'li. Refresh token ham almashadi va
        tanlovni eslab qoladi.
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_id]
              properties:
                company_id:
                  type: [integer, "null"]
                  format: int64
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "401":
          $ref: "#/components/responses/SessionEnded"
        "403":
          description: User bu company'ga a'zo emas (not_member)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /app/me:
    get:
      operationId: getAppMe
      summary: User, tanlangan company va barcha company'lari
      security:
        - appBearer: []
      responses:
        "200":
          description: Joriy user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
  /app/me/nav:
    put:
      operationId: setNavOrder
      summary: Menyu tartibini saqlash
      description: >
        A'zo tanlangan kompaniyadagi bo'limlar tartibini saqlaydi
        (logic/roles.md, 8-bo'lim): kalitlar katalogdan, takror bir marta;
        null standart tartibga qaytaradi. Javob /app/me yangi tartib bilan.
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavOrderInput"
      responses:
        "200":
          description: Joriy user, yangi tartib bilan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/CompanyRequired"
  /app/employees:
    get:
      operationId: listEmployees
      summary: Kompaniya a'zolari (employees.view ruxsati)
      description: >
        Access token'dagi kompaniyaning a'zolari: owner birinchi, keyin
        xodimlar qo'shilgan tartibda. Ism shu kompaniyadagi ism.
      security:
        - appBearer: []
      responses:
        "200":
          description: A'zolar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: addEmployee
      summary: Xodim qo'shish (employees.create ruxsati)
      description: >
        Raqam access token'dagi kompaniyaga user rolida qo'shiladi. Tizimda
        yo'q raqamdan user yaratiladi; boshqa kompaniyada bor raqam shu
        kompaniyaga ham a'zo bo'ladi (multi-user). Javob ikkala holatda bir
        xil: kiritilgan ism qaytadi, raqam oldin tizimda bo'lgan-bo'lmagani
        bilinmaydi. Qo'shilgan odamga xabar yuborilmaydi.
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Qo'shilgan xodim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/AlreadyMember"
  /app/employees/{phone}:
    parameters:
      - name: phone
        in: path
        required: true
        description: Xodimning telefoni, 998901234567 ko'rinishida
        schema:
          type: string
    patch:
      operationId: renameEmployee
      summary: Xodim ismini o'zgartirish (employees.edit ruxsati)
      description: >
        Xodimning shu kompaniyadagi ismi o'zgaradi; boshqa kompaniyalardagi
        ismi o'zgarmaydi. Owner'ning ismi bu yerdan o'zgarmaydi.
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameMember"
      responses:
        "200":
          description: Yangi ismli xodim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/EmployeeNotFound"
        "409":
          $ref: "#/components/responses/CannotChangeOwner"
    delete:
      operationId: removeEmployee
      summary: Xodimni kompaniyadan chiqarish (employees.delete ruxsati)
      description: >
        Xodimning shu kompaniyadagi a'zoligi o'chadi va u keyingi so'rovdayoq
        bu kompaniyaga kira olmaydi. User yozuvi va boshqa kompaniyalardagi
        a'zoliklari qoladi; boshqa kompaniyasi qolmasa, tizimga kira olmaydi.
        Keyin qayta qo'shish mumkin.
      security:
        - appBearer: []
      responses:
        "204":
          description: Xodim chiqarildi
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/EmployeeNotFound"
        "409":
          $ref: "#/components/responses/CannotChangeOwner"
  /app/customer-dropdowns:
    get:
      operationId: listCustomerDropdowns
      summary: Dropdownlar, variantlari bilan
      description: >
        Access token'dagi kompaniyaning dropdownlari yaratilgan tartibda, har
        biri variantlari bilan (o'z tartibida, nofaollari ham). Har a'zo
        o'qiydi: mijoz formasi va ro'yxati shundan quriladi.
      security:
        - appBearer: []
      responses:
        "200":
          description: Dropdownlar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerDropdown"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/CompanyRequired"
    post:
      operationId: createCustomerDropdown
      summary: Dropdown yaratish (settings.create ruxsati)
      description: >
        Variantsiz yangi dropdown. Nomi kompaniyada takrorlanmaydi
        (katta-kichik harf farqsiz), 60 belgigacha.
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameInput"
      responses:
        "201":
          description: Yaratilgan dropdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDropdown"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/CustomerSettingConflict"
  /app/customer-types:
    get:
      operationId: listCustomerTypes
      summary: Mijoz turlari, maydonlari bilan
      description: >
        Access token'dagi kompaniyaning mijoz turlari o'z tartibida, har biri
        maydonlari bilan. Har a'zo o'qiydi: mijoz formasi va ro'yxat ustunlari
        shundan quriladi. Telefon maydon emas: u har mijozda bor.
      security:
        - appBearer: []
      responses:
        "200":
          description: Turlar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerType"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/CompanyRequired"
    post:
      operationId: createCustomerType
      summary: Mijoz turini yaratish (settings.create ruxsati)
      description: >
        Maydonsiz yangi tur, turlar oxiriga. Nomi kompaniyada takrorlanmaydi
        (katta-kichik harf farqsiz), 60 belgigacha.
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameInput"
      responses:
        "201":
          description: Yaratilgan tur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerType"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/CustomerSettingConflict"
  /app/customer-dropdowns/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    patch:
      operationId: renameCustomerDropdown
      summary: Dropdown nomini o'zgartirish (settings.edit ruxsati)
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameInput"
      responses:
        "200":
          description: Yangi nomli dropdown, variantlari bilan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDropdown"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CustomerSettingNotFound"
        "409":
          $ref: "#/components/responses/CustomerSettingConflict"
    delete:
      operationId: deleteCustomerDropdown
      summary: Dropdownni o'chirish (settings.delete ruxsati)
      description: >
        Dropdown yashiriladi (bazadan o'chmaydi), nomi bo'shaydi. Maydonga
        ulangan dropdown o'chirilmaydi: 409 dropdown_in_use.
      security:
        - appBearer: []
      responses:
        "204":
          description: O'chirildi
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CustomerSettingNotFound"
        "409":
          $ref: "#/components/responses/CustomerSettingConflict"
  /app/customer-dropdowns/{id}/options:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    post:
      operationId: addCustomerDropdownOption
      summary: Dropdownga variant qo'shish (settings.create ruxsati)
      description: >
        Variant ro'yxat oxiriga qo'shiladi, faol holda. Nomi dropdown ichida
        takrorlanmaydi (katta-kichik harf farqsiz).
      security:
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptionInput"
      responses:
        "201":
          description: Qo'shilgan variant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOption"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/SubscriptionExpired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CustomerSettingNotFound"
        "409":
          $ref: "#/components/responses/CustomerSettingConflict"
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
    appBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "HS256, aud: app, sub: telefon, company_id va role (tanlangan bo'lsa)"
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
...
  responses:
    BadRequest:
      description: >
        So'rov noto'g'ri (bad_request) yoki maydon xato (validation_error,
        message aniq sababni aytadi)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Juda ko'p so'rov (too_many_requests)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    SignedIn:
      description: Kirildi; refresh token Set-Cookie refresh_token'da
      headers:
        Set-Cookie:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Tokens"
...
    CompanyRequired:
      description: Sessiya hali kompaniya tanlamagan (company_required)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    TelegramLogin:
      type: object
      required: [initData]
      properties:
        initData:
          type: string
          description: window.Telegram.WebApp.initData
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: snake_case xato kodi
        message:
          type: string
          description: O'zbekcha matn
...
    CodeSent:
      type: object
      required: [retry_after]
      properties:
        retry_after:
          type: integer
          description: Keyingi kodni necha soniyadan keyin so'rash mumkin
...